# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.SubmitLinkRequest:
    properties:
      url:
        type: string
      user_id:
        type: string
    required:
    - url
    type: object
  handler.SubmitVideoRequest:
    properties:
      title:
        type: string
      user_id:
        type: string
      video_path:
        type: string
    required:
    - video_path
    type: object
...
  handler.UploadVideoResponse:
    properties:
      file_name:
        type: string
      message:
        type: string
      success:
        type: boolean
      video_id:
        type: string
      video_path:
        type: string
    type: object
...
  handler.VideoProcessData:
    properties:
      audio_path:
        type: string
      processed_at:
        type: string
      status:
        type: string
      transcript:
        $ref: '#/definitions/tools.TranscriptResult'
      video_id:
        type: string
      video_path:
        type: string
    type: object
  handler.VideoProcessResponse:
    properties:
      data:
        $ref: '#/definitions/handler.VideoProcessData'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  tools.TranscriptResult:
    properties:
      full_text:
        type: string
      language:
        type: string
      segments:
        items:
          $ref: '#/definitions/tools.TranscriptSegment'
        type: array
      srt_path:
        description: SRT 字幕文件路径
        type: string
    type: object
  tools.TranscriptSegment:
    properties:
      end:
        type: number
      start:
        type: number
      text:
        type: string
    type: object
host: 127.0.0.1:8096
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for downloading YouTube videos and uploading to Bilibili with AI-powered
    processing
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: YouTube to Bilibili API
  version: "1.0"
paths:
...
  /api/v1/video-process/submit-link:
    post:
      consumes:
      - application/json
      description: 提交YouTube视频链接，系统会自动下载视频、提取音频、转录文本
      parameters:
      - description: 视频链接提交请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.VideoProcessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.VideoProcessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.VideoProcessResponse'
      summary: 提交视频链接进行处理
      tags:
      - video-process
  /api/v1/video-process/submit-video:
    post:
      consumes:
      - application/json
      description: 提交已上传的视频文件路径，系统会提取音频并转录文本
      parameters:
      - description: 视频文件提交请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.VideoProcessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.VideoProcessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.VideoProcessResponse'
      summary: 提交上传的视频进行处理
      tags:
      - video-process
  /api/v1/video-process/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传本地视频文件到服务器
      parameters:
      - description: 视频文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UploadVideoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.UploadVideoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.UploadVideoResponse'
      summary: 上传视频文件
      tags:
      - video-process
...
  /youtube/feed/videos:
    get:
      consumes:
      - application/json
      description: 获取订阅频道的最新视频（支持分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'videos: array, total: int, page: int'
          schema:
            additionalProperties: true
            type: object
      summary: 获取最新视频列表
      tags:
      - youtube
...
  /youtube/video/{videoId}:
    get:
      consumes:
      - application/json
      description: 根据视频ID获取YouTube视频详细信息
      parameters:
      - description: 视频ID
        in: path
        name: videoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'video: object'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: string'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: string'
          schema:
            additionalProperties: true
            type: object
      summary: 获取视频详情
      tags:
      - youtube
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"